+++ packages/kitchensink/app01/feature-hash.js
@@ -1,0 +1,8 @@
+module.exports = {
+  name: 'hash',
+  hook: '$FASTIFY_GET',
+  handler: {
+    url: '/hash/encode/:input',
+    handler: async (request) => request.hash.encode(request.params.input),
+  },
+};