+++ packages/service-fastify-jwt/src/index.js
@@ -1,3 +1,0 @@
-export { default } from './lib/index'
-export * from './lib/index'
-export * from './lib/hooks'