+++ packages/service-hash/src/index.js
@@ -46,7 +46,7 @@ const genSalt = (rounds) =>
     });
   });
 
-const serviceBcrypt = ({ registerAction }) =>
+const serviceBcrypt = ({ registerAction }) => {
   registerAction({
     hook: '$INIT_SERVICES',
     name: hooks.SERVICE_NAME,
@@ -78,6 +78,35 @@ const serviceBcrypt = ({ registerAction }) =>
       };
     },
   });
+
+  // Fastify Integration (optional hook)
+  registerAction({
+    hook: '$FASTIFY_HACKS_AFTER?',
+    name: hooks.SERVICE_NAME,
+    trace: __filename,
+    handler: ({ fastify }, { getContext }) => {
+      const hash = getContext('hash');
+
+      // Prepare the shape of the decorators
+      fastify.decorate('hash', hash);
+      fastify.decorateRequest('hash', null);
+      fastify.decorateReply('hash', null);
+
+      // Add the references using hooks to comply with the decoratos API
+      // https://www.fastify.io/docs/v3.15.x/Decorators/
+
+      fastify.addHook('onRequest', (request, reply, done) => {
+        request.hash = hash;
+        done();
+      });
+
+      fastify.addHook('onResponse', (request, reply, done) => {
+        reply.hash = hash;
+        done();
+      });
+    },
+  });
+};
 
 serviceBcrypt.compare = compare;
 serviceBcrypt.encode = encode;