+++ packages/hooks/CHANGELOG.md
@@ -1,7 +1,8 @@
 # Change Log
 
-## v1.21.0
+## v4.0.0
 
 - Removed babel
 - Features and services names defaults are derived by their registration form and automatically prefixed
 - Added support for positional definition in `registerRoute`
+- Allows for order-independent service2service and feature2feature integration