+++ packages/service-jwt/CHANGELOG.md
@@ -1,8 +1,9 @@
 # Change Log
 
-## v1.21.0
+## v4.0.0
 
 - Remove Babel
+- Integrates with `service-fastify` (conditionally)
 
 ## v3.13.3